+++ react/components/MiniCartContent.js
@@ -1,16 +1,16 @@
 import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
 import { injectIntl, intlShape } from 'react-intl'
-import { reduceBy, values, clone, last, split } from 'ramda'
+import { reduceBy, values, clone, last, split, find, propEq } from 'ramda'
 import classNames from 'classnames'
 import { ExtensionPoint } from 'vtex.render-runtime'
 import { Button, Spinner, IconDelete } from 'vtex.styleguide'
-import ProductPrice from 'vtex.store-components/ProductPrice'
 import { MiniCartPropTypes } from '../propTypes'
 import { toHttps, changeImageUrlSize } from '../utils/urlHelpers'
-import { groupItemsWithParents, isParentItem, isSingleChoiceOption } from '../utils/itemsHelper'
+import { groupItemsWithParents, isSingleChoiceOption } from '../utils/itemsHelper'
 
 import minicart from '../minicart.css'
+import MiniCartFooter from './MiniCartFooter';
 
 /**
  * Minicart content component
@@ -22,7 +22,7 @@ const TOAST_TIMEOUT = 4000
 class MiniCartContent extends Component {
   static propTypes = {
     /* Set the mini cart content size */
-    large: PropTypes.bool,
+    isSizeLarge: PropTypes.bool,
     /* Internationalization */
     intl: intlShape.isRequired,
     /** Define a function that is executed when the item is clicked */
@@ -36,6 +36,7 @@ class MiniCartContent extends Component {
     enableQuantitySelector: MiniCartPropTypes.enableQuantitySelector,
     maxQuantity: MiniCartPropTypes.maxQuantity,
     showDiscount: MiniCartPropTypes.showDiscount,
+    showShippingCost: MiniCartPropTypes.showShippingCost,
   }
 
   state = { isUpdating: [] }
@@ -54,11 +55,14 @@ class MiniCartContent extends Component {
         this.props.data.orderForm.items
       )
     )
+  
+  getShippingCost = orderForm => {
+    const totalizer = find(propEq('id', 'Shipping'))(orderForm.totalizers)
+    return totalizer && totalizer.value / 100
+  }
 
   calculateDiscount = (items, totalPrice) =>
     this.sumItemsPrice(items) - totalPrice
-
-  handleClickButton = () => location.assign('/checkout/#/cart')
 
   handleItemRemoval = async id => {
     this.updateItemLoad(id, true)
@@ -200,7 +204,8 @@ class MiniCartContent extends Component {
     showDiscount,
     actionOnClick,
     isUpdating,
-    large
+    isSizeLarge,
+    showShippingCost
   ) => {
     const items = groupItemsWithParents(orderForm)
     const MIN_ITEMS_TO_SCROLL = 2
@@ -208,28 +213,13 @@ class MiniCartContent extends Component {
     const classes = classNames(
       `${minicart.content} overflow-x-hidden pa1 overflow-y-auto`,
       {
-        [`${minicart.contentSmall} bg-base`]: !large,
-        [`${minicart.contentLarge}`]: large,
-        'overflow-y-scroll': items.length > MIN_ITEMS_TO_SCROLL && !large,
-        'overflow-y-hidden': items.length <= MIN_ITEMS_TO_SCROLL && !large,
+        [`${minicart.contentSmall} bg-base`]: !isSizeLarge,
+        [`${minicart.contentLarge}`]: isSizeLarge,
+        'overflow-y-scroll': items.length > MIN_ITEMS_TO_SCROLL && !isSizeLarge,
+        'overflow-y-hidden': items.length <= MIN_ITEMS_TO_SCROLL && !isSizeLarge,
       }
     )
 
-    const priceAndDiscountClasses = classNames(
-      `${minicart.contentDiscount} w-100 flex justify-end items-center mb3`,
-      {
-        [`pv3`]: large
-      }
-    )
-
-    const checkoutButtonClasses = classNames(
-      ``,
-      {
-        [`bb bw4 bw2-m b--transparent`]: large
-      }
-    )
-
-    const discount = this.calculateDiscount(items, orderForm.value)
     return (
       <Fragment>
         <div className={classes}>
@@ -264,44 +254,16 @@ class MiniCartContent extends Component {
             </Fragment>
           ))}
         </div>
-
-        <div className={`${minicart.contentFooter} w-100 bg-base pa4 bt b--muted-3 pv5 flex flex-column items-end`}>
-          {showDiscount && discount > 0 && (
-            <div className={priceAndDiscountClasses}>
-              <span className="ttl c-action-primary">{labelDiscount}</span>
-              <ProductPrice
-                sellingPriceClass='c-action-primary ph2 dib'
-                sellingPrice={discount}
-                listPrice={discount}
-                showLabels={false}
-                showListPrice={false}
-              />
-            </div>
-          )}
-          <div className={`${minicart.contentPrice} mb3`}>
-            {this.isUpdating
-              ? (<Spinner size={18} />)
-              : (
-                <ProductPrice
-                  sellingPriceClass='c-muted-1 b ph2 dib'
-                  sellingPrice={orderForm.value}
-                  listPrice={orderForm.value}
-                  showLabels={false}
-                  showListPrice={false}
-                />
-              )
-            }
-          </div>
-          <div className={checkoutButtonClasses}>
-            <Button
-              variation="primary"
-              size="small"
-              onClick={this.handleClickButton}
-              >
-              {label}
-            </Button>
-          </div>
-        </div>
+        <MiniCartFooter
+          shippingCost={this.getShippingCost(orderForm)}
+          isUpdating={this.isUpdating}
+          totalValue={orderForm.value}
+          buttonLabel={label}
+          isSizeLarge={isSizeLarge}
+          labelDiscount={labelDiscount}
+          showDiscount={showDiscount}
+          showShippingCost={showShippingCost}
+        />
       </Fragment>
     )
   }
@@ -320,7 +282,8 @@ class MiniCartContent extends Component {
       intl,
       showDiscount,
       actionOnClick,
-      large,
+      isSizeLarge,
+      showShippingCost,
     } = this.props
     const { isUpdating } = this.state
 
@@ -347,7 +310,8 @@ class MiniCartContent extends Component {
       showDiscount,
       actionOnClick,
       isUpdating,
-      large
+      isSizeLarge,
+      showShippingCost
     )
   }
 }